perf(sqlite): remove safe mode version preflight
